more cleanup of rats.cli
