+++ rats-apps/src/python/rats/cli/__init__.py
@@ -1,19 +1,17 @@
 """Uses `rats.annotations` to streamline the creation of CLI commands written with Click."""
 
-from ._annotations import CommandContainer, CommandId, command, group
+from ._annotations import CommandId, command, group
 from ._click import ClickCommandGroup, ClickCommandMapper
-from ._executable import ClickExecutable, ClickGroup
-from ._plugins import AttachClickCommands, ClickGroupPlugin
+from ._executable import ClickExecutable
+from ._plugins import AttachClickCommands, AttachClickGroup, ClickGroupPlugin, CommandContainer
 
 __all__ = [
     "CommandId",
     "command",
     "group",
-    "CommandContainer",
     "ClickCommandMapper",
     "ClickExecutable",
     "ClickGroupPlugin",
     "ClickCommandGroup",
     "AttachClickCommands",
-    "ClickGroup",
 ]
