+++ rats-apps/src/python/rats/cli/_annotations.py
@@ -1,14 +1,9 @@
 from __future__ import annotations
 
 from collections.abc import Callable
-from functools import partial
-from typing import Any, NamedTuple, TypeVar
-
-import click
+from typing import NamedTuple, TypeVar
 
 from rats import annotations
-
-from ._plugins import ClickGroupPlugin
 
 
 class CommandId(NamedTuple):
@@ -45,32 +40,3 @@ def group(command_id: CommandId) -> annotations.DecoratorType:
 
 def get_class_commands(cls: type) -> annotations.AnnotationsContainer:
     return annotations.get_class_annotations(cls).with_namespace("commands")
-
-
-class CommandContainer(ClickGroupPlugin):
-    def on_group_open(self, group: click.Group) -> None:
-        def cb(_method: Callable[[Any], Any], *args: Any, **kwargs: Any) -> None:
-            """
-            Callback handed to `click.Command`. Calls the method with matching name on this class.
-
-            When the command is decorated with `@click.params` and `@click.option`, `click` will
-            call this callback with the parameters in the order they were defined. This callback
-            then calls the method with the same name on this class, passing the parameters in
-            reverse order. This is because the method is defined with the parameters in the
-            reverse order to the decorator, so we need to reverse them again to get the correct
-            order.
-            """
-            _method(*args, **kwargs)
-
-        for tate in get_class_commands(type(self)).annotations:
-            method = getattr(self, tate.name)
-            params = list(reversed(getattr(method, "__click_params__", [])))
-            for command in tate.groups:
-                group.add_command(
-                    click.Command(
-                        name=command.name,
-                        callback=partial(cb, method),
-                        short_help=method.__doc__,
-                        params=params,
-                    )
-                )
